requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.6
django-cors-headers==3.6.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
